feat(free-trial): move logic to a context, avoid refetching on every mount of the sidebar
